WoT restructure: nest boxLayout/ under viewport/ (viewport is its only consumer)
