+++ handsontable/src/3rdparty/walkontable/src/viewport/boxLayout/layoutSnapshot.ts
