+++ handsontable/src/3rdparty/walkontable/src/viewport/boxLayout/resolveLayout.ts
